+++ commands/history/clear-message-history.js
@@ -1,4 +1,4 @@
-const mysqlQuery = require('../../tools/mysql-query');
+const mysqlQuery = require('../../tools/mysqlQuery');
 module.exports = {
     name: 'clear-message-history',
     description: "Clears all message history from db",
