+++ tools/captureMessage.js
